pyproject.toml
[tool.poetry]
name = "relia_oss"
version = "1.0.0"
description = "Cloud cost prevention and optimization tool."
authors = ["Relia Team"]
readme = "README.md"
homepage = "https://github.com/davidahmann/relia_oss"
repository = "https://github.com/davidahmann/relia_oss"
keywords = ["finops", "terraform", "cost", "cloud", "aws"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Utilities",
]
packages = [{include = "relia"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.20.0"
rich = "^13.0.0"
pydantic = "^2.0.0"
python-hcl2 = "^7.3.1"
boto3 = "^1.42.4"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.1.0"
mypy = "^1.0.0"
pre-commit = "^3.5.0"
bandit = "^1.7.5"
pytest-cov = "^7.0.0"

[tool.pytest.ini_options]
addopts = "--cov=relia --cov-fail-under=85"
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests"]
targets = ["relia"]
skips = ["B301"] # Skip pickle warning if present (or others)

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
relia = "relia.main:app"
